CIV-1853 refactor calculate-amount.js
