+++ src/main/routes/calculateMonthlyIncomeExpense/monthlyExpenseIncomeCalculatorController.ts
@@ -1,6 +1,6 @@
 import * as express from 'express';
 import {TOTAL_AMOUNT_CALCULATION_URL} from '../urls';
-import {calculateTotalAmount} from '../../common/utils/calculateMonthlyIncomeExpeses/monthlyIncomeExpensesCalculator';
+import {calculateTotalAmount} from '../../common/utils/calculateMonthlyIncomeExpenses/monthlyIncomeExpensesCalculator';
 
 const monthlyExpenseIncomeCalculatorController = express.Router();
